run_answer_tests.sh: run make once at the start
This ensures any compilation errors only appear once at the beginning of the
output, rather than interleaved with the test results. It also speeds up the
test runs slightly, as `make -q` takes around 50ms on xrb (mostly due to
`-include $(DEP)`). If a target fails to compile, its tests will be skipped,
but still be counted as failing for the exit code.