2023: day 21 part 2 wip
This approach works, but scales quite poorly. Most of the time is spent
in next_edge_set.emplace(), and we can probably replace the maps with
vectors if we're careful to avoid duplicates.

For valid inputs, it should also be possible to skip over the
fully-reachable tiles entirely and only simulate the last few
iterations.