2023 day 21: reuse per-tile reachability calculations
Based on the assumptions we make about the input, each off-axis tile in
a given direction and iteration will be identical (which is confirmed by
the `reachable_cache` statistics). Instead of using a cache, just
calculate the number of reachable plots and multiply by the number of
tiles on the edge.